requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
langchain-core==0.1.16
langsmith==0.0.83
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20 